refactor: reshuffle order book domain for extensibility
